Reorganize all the tests
