--- cuda_bindings/cuda/bindings/_test_helpers/__init__.py
@@ -1,6 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-# This package contains test helper utilities that may also be useful for other libraries outside of `cuda.bindings`,
-# such as `cuda.core`. These utilities are not part of the public API of `cuda.bindings` and may change without notice.
