+++ cuda_bindings/tests/nvml/conftest.py
@@ -4,9 +4,9 @@
 from collections import namedtuple
 
 import pytest
+from cuda_python_test_helpers.arch_check import unsupported_before  # noqa: F401
 
 from cuda.bindings import nvml
-from cuda.bindings._test_helpers.arch_check import unsupported_before  # noqa: F401
 
 
 class NVMLInitializer:
