+++ cuda_bindings/tests/test_examples.py
@@ -7,8 +7,7 @@ import subprocess
 import sys
 
 import pytest
-
-from cuda.bindings._test_helpers.pep723 import has_package_requirements_or_skip
+from cuda_python_test_helpers.pep723 import has_package_requirements_or_skip
 
 examples_path = os.path.join(os.path.dirname(__file__), "..", "examples")
 examples_files = glob.glob(os.path.join(examples_path, "**/*.py"), recursive=True)
