+++ cuda_core/tests/graph/test_graph_definition.py
@@ -7,10 +7,10 @@ from collections.abc import Callable
 from dataclasses import dataclass, field
 
 import pytest
+from conftest import xfail_on_graph_mempool_oom
 from helpers.graph_kernels import compile_common_kernels
 from helpers.misc import try_create_condition
 
-from conftest import xfail_on_graph_mempool_oom
 from cuda.core import Device, LaunchConfig
 from cuda.core.graph import (
     AllocNode,
