+++ cuda_core/tests/graph/test_graph_definition_lifetime.py
@@ -13,10 +13,10 @@ import time
 import weakref
 
 import pytest
+from conftest import xfail_on_graph_mempool_oom
 from helpers.graph_kernels import compile_common_kernels
 from helpers.misc import try_create_condition
 
-from conftest import xfail_on_graph_mempool_oom
 from cuda_python_test_helpers import under_compute_sanitizer
 
 # Resource finalization triggered by graph destruction is not synchronous. A
