+++ cuda_core/tests/graph/test_graph_memory_resource.py
@@ -5,10 +5,10 @@
 """Tests for GraphMemoryResource allocation and attributes during graph capture."""
 
 import pytest
+from conftest import xfail_on_graph_mempool_oom
 from helpers import IS_WINDOWS, IS_WSL
 from helpers.buffers import compare_buffer_to_constant, make_scratch_buffer, set_buffer
 
-from conftest import xfail_on_graph_mempool_oom
 from cuda.core import (
     Device,
     DeviceMemoryResource,
