+++ cuda_core/tests/memory/test_managed_ops.py
@@ -4,9 +4,9 @@
 import mmap
 
 import pytest
+from conftest import create_managed_memory_resource_or_skip
 from helpers.buffers import DummyDeviceMemoryResource, DummyUnifiedMemoryResource
 
-from conftest import create_managed_memory_resource_or_skip
 from cuda.bindings import driver
 from cuda.core import Device, Host, ManagedBuffer
 from cuda.core._memory._managed_buffer import _get_int_attr
