+++ cuda_core/tests/test_object_protocols.py
@@ -13,10 +13,10 @@ import re
 import weakref
 
 import pytest
+from conftest import xfail_on_graph_mempool_oom
 from helpers.graph_kernels import compile_common_kernels
 from helpers.misc import try_create_condition
 
-from conftest import xfail_on_graph_mempool_oom
 from cuda.core import (
     Buffer,
     Device,
