+++ cuda_core/tests/test_tensor_map.py
@@ -3,8 +3,8 @@
 
 import numpy as np
 import pytest
-
 from conftest import create_managed_memory_resource_or_skip, skip_if_managed_memory_unsupported
+
 from cuda.core import (
     Device,
     ManagedMemoryResourceOptions,
