+++ cuda_core/tests/test_utils.py
@@ -26,7 +26,7 @@ except ImportError:
     ml_dtypes = None
 import numpy as np
 import pytest
-from helpers.marks import requires_module
+from cuda_python_test_helpers.marks import requires_module
 
 from cuda.core import Device
 from cuda.core._dlpack import DLDeviceType
